requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
av==10.0.0
certifi @ file:///croot/certifi_1683875369620/work/certifi
charset-normalizer==2.1.1
click==8.1.4
cmake==3.25.0
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.2
fastapi==0.100.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gradio==3.36.1
gradio_client==0.2.8
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lazy_loader==0.3
linkify-it-py==2.0.2
lit==15.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
MedPy==0.4.0
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.23.1
opencv-python==4.8.0.74
orjson==3.9.2
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
pydantic==2.0.2
pydantic_core==2.1.2
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.29.1
requests==2.28.1
rpds-py==0.8.10
scikit-image==0.21.0
scipy==1.11.1
semantic-version==2.10.0
SimpleITK==2.2.1
six==1.16.0
sniffio==1.3.0
spatial-correlation-sampler==0.4.0
starlette==0.27.0
sympy==1.11.1
tifffile==2023.7.10
toolz==0.12.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.13
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
